[auto-doc] improve auto doc with roles/permissions/public information on handlers and related input
