+++ kasper-documentation/src/main/java/com/viadeo/kasper/doc/element/AbstractPropertyDomainElement.java
@@ -9,9 +9,6 @@ package com.viadeo.kasper.doc.element;
 import com.google.common.collect.Lists;
 import com.viadeo.kasper.annotation.XKasperAlias;
 import com.viadeo.kasper.doc.nodes.DocumentedBean;
-import com.viadeo.kasper.security.annotation.XKasperPublic;
-import com.viadeo.kasper.security.annotation.XKasperRequirePermissions;
-import com.viadeo.kasper.security.annotation.XKasperRequireRoles;
 
 import java.util.List;
 
@@ -20,10 +17,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
 public abstract class AbstractPropertyDomainElement extends AbstractDomainElement {
 
     private final DocumentedBean properties;
-    private final boolean publicAccess;
     private final List<String> aliases;
-    private final List<String> roles;
-    private final List<String> permissions;
 
     // ------------------------------------------------------------------------
 
@@ -33,36 +27,6 @@ public abstract class AbstractPropertyDomainElement extends AbstractDomainElemen
         super(checkNotNull(domain), checkNotNull(type), checkNotNull(referenceClass));
 
         this.properties = new DocumentedBean(referenceClass);
-        this.publicAccess = referenceClass.getAnnotation(XKasperPublic.class) != null;
-
-
-        // @XKasperRequireRoles
-        if (null != referenceClass.getAnnotation(XKasperRequireRoles.class)) {
-            final XKasperRequireRoles requireRoles = (XKasperRequireRoles)
-                    referenceClass.getAnnotation(XKasperRequireRoles.class);
-
-            if (null != requireRoles.value()) {
-                this.roles = Lists.newArrayList(requireRoles.value());
-            } else {
-                this.roles = Lists.newArrayList();
-            }
-        } else {
-            this.roles = Lists.newArrayList();
-        }
-
-        // @XKasperRequirePermissions
-        if (null != referenceClass.getAnnotation(XKasperRequirePermissions.class)) {
-            final XKasperRequirePermissions requirePermissions = (XKasperRequirePermissions)
-                    referenceClass.getAnnotation(XKasperRequirePermissions.class);
-
-            if (null != requirePermissions.value()) {
-                this.permissions = Lists.newArrayList(requirePermissions.value());
-            } else {
-                this.permissions = Lists.newArrayList();
-            }
-        } else {
-            this.permissions = Lists.newArrayList();
-        }
 
         final XKasperAlias annotation = (XKasperAlias) referenceClass.getAnnotation(XKasperAlias.class);
         if (null != annotation) {
@@ -78,16 +42,8 @@ public abstract class AbstractPropertyDomainElement extends AbstractDomainElemen
         return properties;
     }
 
-    public boolean isPublicAccess() {
-        return publicAccess;
-    }
-
     public List<String> getAliases() {
         return aliases;
     }
-
-    public List<String> getRoles() { return roles; }
-
-    public List<String> getPermissions() { return permissions; }
 
 }
